feat: add heapsort functionality and heapsort tests
Created two types of heapsort. One that do not sort in-place called 'NaiveSort' and one called 'HeapSort' that sorts the array in place.

The in-place HeapSort invalidates the heap structure. Therefore, I invalidate any further operations on that heap whenever the user has used the Sort() method. To re-instantiate the heap, the user must call BuildHeap method.